+++ lib/rails-bootstrap-helpers/renderers/row_link_renderer.rb
@@ -1,0 +1,12 @@
+module RailsBootstrapHelpers::Renderers
+  class RowLinkRenderer < AbstractButtonRenderer
+    def initialize (template, *args, &block)
+      super template, :link, *args, &block
+    end
+
+    def render
+      append_class "rowlink"
+      super
+    end
+  end
+end